+++ resolwe/flow/views/mixins.py
@@ -7,6 +7,7 @@ from rest_framework.exceptions import ParseError
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
+from resolwe.observers.protocol import post_permission_changed
 from resolwe.permissions.models import get_anonymous_user
 from resolwe.permissions.utils import assign_contributor_permissions
 
@@ -45,10 +46,17 @@ class ResolweCreateModelMixin(mixins.CreateModelMixin):
         """Create a resource."""
         with transaction.atomic():
             instance = serializer.save()
-
-            # Assign all permissions to the object contributor.
-            if hasattr(instance, "permission_group") and not instance.in_container():
-                assign_contributor_permissions(instance)
+            if hasattr(instance, "permission_group"):
+                # Assign all permissions to the object contributor when object is not
+                # in container.
+                if not instance.in_container():
+                    assign_contributor_permissions(instance)
+                # The object was created inheriting permissions from its container.
+                # Notify observers.
+                else:
+                    post_permission_changed.send(
+                        sender=type(instance), instance=instance
+                    )
 
 
 class ResolweUpdateModelMixin(mixins.UpdateModelMixin):